Refactor(api): api 코드 폴더구조 변경 및 경로 수정
