+++ src/api/hooks/Theme/useGetThemesProducts.ts
@@ -1,4 +1,4 @@
-import { fetchInstance } from '../instance';
+import { fetchInstance } from '../../instance';
 import type { GoodsData } from '@/types';
 import {
   type InfiniteData,
